+++ packages/eslint-config-ckeditor5/.eslintrc.js
@@ -8,7 +8,7 @@
 module.exports = {
 	extends: 'eslint:recommended',
 	parserOptions: {
-		ecmaVersion: 6,
+		ecmaVersion: 2018,
 		sourceType: 'module'
 	},
 	env: {
@@ -230,7 +230,11 @@ module.exports = {
 		],
 		'space-before-function-paren': [
 			'error',
-			'never'
+			{
+				anonymous: 'never',
+				named: 'never',
+				asyncArrow: 'always'
+			}
 		],
 		'space-infix-ops': 'error',
 		'space-in-parens': [