+++ packages/jsdoc-plugins/lib/relation-fixer/addmissingdoclets.js
@@ -5,38 +5,26 @@
 
 'use strict';
 
-const { cloneDeep } = require( 'lodash' );
 const getMissingDocletsData = require( './getmissingdocletsdata' );
 const DocletCollection = require( '../utils/doclet-collection' );
 
 module.exports = addMissingDoclets;
 
 /**
- * Adds missing doclets for members coming from implemented interfaces, extended classes, and mixins.
+ * Adds missing doclets for members coming from implemented interfaces, extended classes, mixins and typedefs.
  * It does also support inheriting static members and typedef inheritance, which both are not supported by the JSDoc.
- * This module requires the input preprocessed by the `buildRelations()` function.
+ * This function requires input to be preprocessed by the `buildRelations()` function.
  *
  * @param {Array.<Doclet>} doclets
  * @returns {Array.<Doclet>}
  */
 function addMissingDoclets( doclets ) {
-	doclets = cloneDeep( doclets );
-
 	const docletCollection = new DocletCollection();
-
-	/** @type {Array.<Doclet>} */
-	const typedefDoclets = [];
 
 	for ( const doclet of doclets ) {
 		// Group doclets by memberof property.
 		docletCollection.add( `memberof:${ doclet.memberof }`, doclet );
-
-		if ( doclet.kind === 'typedef' ) {
-			typedefDoclets.push( doclet );
-		}
 	}
-
-	extendTypedefs( typedefDoclets );
 
 	const extensibleDoclets = doclets.filter( doclet => {
 		return (
@@ -106,27 +94,4 @@ function addMissingDoclets( doclets ) {
 		...doclets,
 		...newDocletsToAdd
 	];
-}
-
-/**
- * Copy properties from parent typedefs to typedefs which extend them.
- *
- * @param {Array.<Doclet>} typedefDoclets
- */
-function extendTypedefs( typedefDoclets ) {
-	for ( const typedefDoclet of typedefDoclets ) {
-		for ( const parentLongname of typedefDoclet.augmentsNested ) {
-			const parentDoclet = typedefDoclets.find( doclet => doclet.longname === parentLongname );
-
-			if ( parentDoclet && parentDoclet.properties ) {
-				parentDoclet.properties.forEach( parentProperty => {
-					if ( typedefDoclet.properties && !typedefDoclet.properties.find( p => p.name === parentProperty.name ) ) {
-						const inheritedProperty = cloneDeep( parentProperty );
-						inheritedProperty.inherited = true;
-						typedefDoclet.properties.push( inheritedProperty );
-					}
-				} );
-			}
-		}
-	}
 }