+++ packages/ckeditor5-dev-env/CHANGELOG.md
@@ -1,6 +1,76 @@
 Changelog
 =========
 
+## [15.0.3](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@15.0.2...@ckeditor/ckeditor5-dev-env@15.0.3) (2019-07-23)
+
+### Other changes
+
+* Adjusted repository, documentation and scripts to changes done in mgit/mrgit. See: [cksource/mgit2#85](https://github.com/cksource/mgit2/issues/85). ([8b4d08d](https://github.com/ckeditor/ckeditor5-dev/commit/8b4d08d))
+
+
+## [15.0.2](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@15.0.1...@ckeditor/ckeditor5-dev-env@15.0.2) (2019-07-16)
+
+### Bug fixes
+
+* Adjusted usage of GitHub API inside createGithubRelease() util. Closes [ckeditor/ckeditor5#1889](https://github.com/ckeditor/ckeditor5/issues/1889). ([d9befaf](https://github.com/ckeditor/ckeditor5-dev/commit/d9befaf))
+
+
+## [15.0.1](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@15.0.0...@ckeditor/ckeditor5-dev-env@15.0.1) (2019-07-16)
+
+### Bug fixes
+
+* Adjusted usage of GitHub API to the latest version. All API's methods return an instance of Promise instead of requiring a callback as the last parameter. Closes [ckeditor/ckeditor5#1889](https://github.com/ckeditor/ckeditor5/issues/1889). ([486f732](https://github.com/ckeditor/ckeditor5-dev/commit/486f732))
+
+
+## [15.0.0](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@14.1.2...@ckeditor/ckeditor5-dev-env@15.0.0) (2019-07-15)
+
+### Features
+
+* Before starting the publishing process on NPM, the task will check whether the current user is logged to npm. Removed `tasks.releaseRepository()`. It's now possible to release a single repo with the multi-repo release task. Closes [#496](https://github.com/ckeditor/ckeditor5-dev/issues/496). Closes [#498](https://github.com/ckeditor/ckeditor5-dev/issues/498). ([90dc7a1](https://github.com/ckeditor/ckeditor5-dev/commit/90dc7a1))
+
+  In order to release a single repository, first, you need to bump its version. Then you will be able to publish changes. For bumping version, you can use the following code:
+
+  ```js
+  require( '@ckeditor/ckeditor5-dev-env' )
+  	.bumpVersions( {
+  		cwd: process.cwd(),
+  		packages: null, // <-- It means that you want to call the script for single package.
+  		dryRun: process.argv.includes( '--dry-run' )
+  	} );
+  ```
+
+  For publishing changes:
+
+  ```js
+  require( '@ckeditor/ckeditor5-dev-env' )
+  	.releaseSubRepositories( {
+  		cwd: process.cwd(),
+  		packages: null, // <-- It means that you want to call the script for single package.
+  		dryRun: process.argv.includes( '--dry-run' )
+  	} );
+  ```
+
+  You can call both scripts with `--dry-run` option that allows you testing the entire process. Those scripts won't publish anything in this mode.
+
+### BREAKING CHANGES
+
+* `tasks.releaseRepository()` has been removed. Use `tasks.bumpVersions()` and `tasks.releaseSubRepositories()` instead. 
+
+
+## [14.1.2](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@14.1.1...@ckeditor/ckeditor5-dev-env@14.1.2) (2019-07-15)
+
+### Other changes
+
+* Upgraded dependencies for most of the packages. Merged Lerna + Yarn and they can work together now. Closes [#527](https://github.com/ckeditor/ckeditor5-dev/issues/527). Closes [#466](https://github.com/ckeditor/ckeditor5-dev/issues/466). ([dcc3215](https://github.com/ckeditor/ckeditor5-dev/commit/dcc3215))
+
+
+## [14.1.1](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@14.1.0...@ckeditor/ckeditor5-dev-env@14.1.1) (2019-04-04)
+
+### Bug fixes
+
+* Release tool will not crash when publishing a package for the first time. Closes [#491](https://github.com/ckeditor/ckeditor5-dev/issues/491). ([fcc875d](https://github.com/ckeditor/ckeditor5-dev/commit/fcc875d))
+
+
 ## [14.1.0](https://github.com/ckeditor/ckeditor5-dev/compare/@ckeditor/ckeditor5-dev-env@14.0.1...@ckeditor/ckeditor5-dev-env@14.1.0) (2019-03-28)
 
 ### Features